chore: Refactor Dockerfile and update run script
This commit refactors the Dockerfile for clarity, efficiency, and
improved layer caching.  It streamlines the installation process
for Poetry and project dependencies.

The docker_run_poetry.sh script is updated to handle default host/port
values and uses `exec` to replace the shell process with gunicorn.

A .dockerignore file is added to exclude unnecessary files.